backend/requirements.txt
asgiref==3.5.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
decouple==0.0.7
Deprecated==1.2.13
dj-database-url==1.2.0
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-oauth-toolkit==2.2.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-recursive==0.1.2
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.4.2
MarkupSafe==2.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==22.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyrsistent==0.19.3
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
sqlparse==0.4.3
swagger-spec-validator==3.0.3
tomli==2.0.1
typing_extensions==4.4.0
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.13
user-agents==2.2.0
Werkzeug==2.2.2
whitenoise==6.3.0
wrapt==1.14.1
